% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\name{categories}
\alias{categories}
\title{Categories of board games}
\usage{
categories()
}
\value{
a data frame containing all categories of board games.
}
\description{
Categories of board games
}
\examples{
bg_categories <- categories()
bg_categories
}
\seealso{
[mechanics()]
}
