% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{collection}
\alias{collection}
\title{User's Collection}
\usage{
collection(username)
}
\arguments{
\item{username}{name of the user to retrieve their collection}
}
\value{
Data frame containing the specified user's collection, including those that
they own or want to buy/play.
}
\description{
\code{collection()} retrieves the board game collection for a
  specified user from \href{http://boardgamegeek.com}{Board Game Geek}.
}
\examples{
my_collection <- collection('alizat')
my_collection
}
