% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_to_na.R
\name{empty_to_na}
\alias{empty_to_na}
\title{Convert Array to NA if Empty}
\usage{
empty_to_na(x)
}
\arguments{
\item{x}{an array}
}
\value{
\code{NA} if x was empty. Otherwise, x is returned as is.
}
\description{
\code{empty_to_na()} converts the received array to single value
  of \code{NA} if it is an empty array. Otherwise, the received array is
  returned as is.
}
\examples{
empty_to_na(character(0))
empty_to_na(c())
empty_to_na(c(1,2,3))
empty_to_na(c('a','b','c'))
}
