% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{families}
\alias{families}
\title{Families of Board Games}
\usage{
families(wait = 10, verbose = FALSE)
}
\arguments{
\item{wait}{number of seconds to wait between pages while scraping to avoid
being blocked by BGG (default is 10 seconds)}

\item{verbose}{whether to print supplementary text (shows intermediate
progress)}
}
\value{
Data frame containing list of families from BGG website
}
\description{
Families of Board Games
}
\examples{
families(verbose = TRUE)
}
\seealso{
\code{\link{categories}} \code{\link{mechanics}}
  \code{\link{designers}}
}
