% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{family}
\alias{family}
\title{Family details}
\usage{
family(family_id, family_type = "boardgamefamily")
}
\arguments{
\item{family_id}{id of the desired family.}

\item{family_type}{filter results by the specified family type (if any).
Valid values are \code{"boardgamefamily"} (default), \code{"rpg"} and
\code{"rpgperiodical"}.}
}
\value{
Data frame containing the details for the specified family.
}
\description{
\code{family()} retrieves the details for the specified family.
}
\examples{
family(78462)
family(43631)
}
