% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_extractor.R
\name{features_extractor}
\alias{features_extractor}
\title{Extract features from HTML elements}
\usage{
features_extractor(elements, features)
}
\arguments{
\item{elements}{a list of HTML elements}

\item{features}{a list of features to obtain for each of the HTML elements
that are given as input}
}
\value{
A data frame containing the retrieved features for the HTML elements that are
given as input
}
\description{
\code{features_extractor()} obtains, for a list of HTML elements,
  a set of features that are given as input.
}
\examples{
my_html <- rvest::read_html('https://boardgamegeek.com/xmlapi2/collection?username=alizat')
my_elements <- rvest::html_elements(my_html, 'item')
my_features <- c('name', 'status::own', 'status::wanttoplay', 'status::wanttobuy')
features_extractor(my_elements, my_features)
}
