% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum.R
\name{forum}
\alias{forum}
\title{List of Threads in a Forum}
\usage{
forum(forum_id)
}
\arguments{
\item{forum_id}{id of the desired forum.}
}
\value{
Data frame containing list of threads for the specified forum.
}
\description{
\code{forum()} retrieves the list of threads for the specified
  forum.
}
\examples{
forum(2418)
}
\seealso{
\code{\link{forumlist}} \code{\link{thread}}
}
