% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forumlist.R
\name{forumlist}
\alias{forumlist}
\title{List of Forums for a Specified Item}
\usage{
forumlist(forumlist_id, type = "thing")
}
\arguments{
\item{type}{type of the specified item. Valid values are \code{"thing"} and
\code{"family"}.}

\item{id}{id of the item that you wish retrieve the forum lists for.}
}
\value{
Data frame containing list of available forums for the specified item.
}
\description{
\code{forumlist()} retrieves the lists of forums available for
  the specified item.
}
\examples{
forumlist_pandemic_on_the_brink <- forumlist(40849)
forumlist_pandemic_on_the_brink
}
\seealso{
\code{\link{forum}} \code{\link{thread}}
}
