% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guild.R
\name{guild}
\alias{guild}
\title{Guild Info}
\usage{
guild(guild_id, members = 1, sort_by = "username")
}
\arguments{
\item{guild_id}{id of the guild you wish to retrieve}

\item{members}{include member roster in the results? (default: \code{TRUE})}

\item{sort_by}{how to sort the members list. Valid values are
\code{"username"} (default) and \code{"date"}.}
}
\value{
List containing two data frames: \code{"guild_details"} and
\code{"members_info"}.
}
\description{
\code{guild()} retrieves the info for the specified guild
  (details and members list).
}
\examples{
guild(1299)
}
