% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot.R
\name{hot}
\alias{hot}
\title{Hot items nowadays}
\usage{
hot(type = "boardgame")
}
\arguments{
\item{type}{type of hot items to return. Possible values are
\code{"boardgame"} (default), \code{"rpg"}, \code{"videogame"},
\code{"boardgameperson"}, \code{"rpgperson"}, \code{"boardgamecompany"},
\code{"rpgcompany"} and \code{"videogamecompany"}.}
}
\value{
A data frame containing nowadays' hot items.
}
\description{
\code{hot()} retrieves the items that are "hot" nowadays.
}
\examples{
hot_items_df <- hot()
hot_items_df
}
