% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanics.R
\name{mechanics}
\alias{mechanics}
\title{Mechanics of board games}
\usage{
mechanics()
}
\value{
A data frame containing all mechanics of board games.
}
\description{
\code{mechanics()} retrieves the board game mechanics that are
  present at \href{http://boardgamegeek.com}{Board Game Geek}.
}
\examples{
bg_mechanics <- mechanics()
bg_mechanics
}
\seealso{
\code{\link{categories}}
}
