% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plays.R
\name{plays}
\alias{plays}
\title{Plays of a specific item}
\usage{
plays(
  item_id = NULL,
  username = NULL,
  type = NULL,
  mindate = NULL,
  maxdate = NULL,
  subtype = NULL,
  wait = 5
)
}
\arguments{
\item{item_id}{id of item to retrieve plays information on. At least one of
\code{item_id} and \code{username} needs to be supplied.}

\item{username}{username of user to retrieve plays for. At least one of
\code{item_id} and \code{username} needs to be supplied.}

\item{type}{type of the item you want to request play information for. Valid
values are \code{"thing"} and \code{"family"}.}

\item{mindate}{if supplied, returns only plays of specified date or later.}

\item{maxdate}{if supplied, returns only plays of specified date or previous.}

\item{subtype}{filters plays by supplied subtype. Valid values are
\code{"boardgame"} (default), \code{"boardgameexpansion"},
\code{"boardgameaccessory"}, \code{"boardgameintegration"},
\code{"boardgamecompilation"}, \code{"boardgameimplementation"},
\code{"rpg"}, \code{"rpgitem"} and \code{"videogame"}}

\item{wait}{number of seconds to wait between pages while scraping to avoid
being blocked by BGG (default is 5 seconds)}
}
\value{
Plays info of supplied item and/or username
}
\description{
\code{plays()} retrieves info on either plays of a specific
  item, plays by a specific user or plays of a specific item by a specific
  user.
}
\examples{
plays_3_wishes <- plays(198836)
plays_3_wishes
}
