% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchbgg.R
\name{searchbgg}
\alias{searchbgg}
\title{BGG Search}
\usage{
searchbgg(query, type = NULL, exact = 0)
}
\arguments{
\item{query}{keywords of item that you are looking for}

\item{type}{type of item. Possible values are \code{"rpgitem"},
\code{"videogame"}, \code{"boardgame"}, \code{"boardgameaccessory"} and
\code{"boardgameexpansion"}. Specifying multiple types separated by commas
is allowed (see examples below).}

\item{exact}{whether the returned result should match the search query
perfectly (default is \code{0})}
}
\value{
Data frame containing search results.
}
\description{
\code{searchbgg()} gets search results for supplied query.
}
\examples{
searchbgg(query = 'shipshape')
searchbgg(query = 'shipshape', type = 'boardgame')
searchbgg(query = 'dune imperium')
searchbgg(query = 'water', type = 'rpgitem')
}
