% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thing.R
\name{thing}
\alias{thing}
\title{Thing Info}
\usage{
thing(ids)
}
\arguments{
\item{ids}{IDs of the items that you wish retrieve the information for.}
}
\value{
Data frame containing the details of the items for which the IDs were
supplied.
}
\description{
\code{thing()} retrieves the available information of the
  specified items (i.e. board games and board game expansions).
}
\examples{
games_details <- thing(c(158600, 194607, 40849))
games_details
}
