% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread.R
\name{thread}
\alias{thread}
\title{Thread Details}
\usage{
thread(thread_id)
}
\arguments{
\item{thread_id}{id of the desired forum.}
}
\value{
Data frame containing the details of the specified thread.
}
\description{
\code{thread()} retrieves the details for the specified thread.
}
\examples{
pandemic_on_the_brink_review <- thread(650169)
pandemic_on_the_brink_review
}
\seealso{
\code{\link{forumlist}} \code{\link{forum}}
}
