% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user}
\alias{user}
\title{User Info}
\usage{
user(username, buddies = 0, guilds = 0, hot = 0, top = 0)
}
\arguments{
\item{username}{result to include user's top items?}
}
\value{
List containing the specified user's details.
}
\description{
\code{user()} retrieves the details for a specified user from
  \href{http://boardgamegeek.com}{Board Game Geek}.
}
\examples{
user_info <- user('alizat')
user_info
}
\seealso{
\code{\link{collection}}
}
