% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\name{categories}
\alias{categories}
\title{Categories of board games}
\usage{
categories()
}
\value{
A data frame containing all categories of board games (as per BGG).
}
\description{
\code{categories()} retrieves the board game categories that are
  present at \href{http://boardgamegeek.com}{Board Game Geek}.
}
\examples{
bg_categories <- categories()
bg_categories
}
\seealso{
\code{\link{mechanics}}
}
