% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{collection}
\alias{collection}
\title{User's Collection}
\usage{
collection(username)
}
\arguments{
\item{username}{name of the user to retrieve their collection}
}
\value{
Collection of games for the specified user, include those that they
own or want to buy/play.
}
\description{
User's Collection
}
\examples{
my_collection <- collection('alizat')
my_collection
}
