% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot.R
\name{hot}
\alias{hot}
\title{Hot items nowadays}
\usage{
hot(type = "boardgame")
}
\arguments{
\item{type}{type of hot items to return. Possible values are `boardgame`
(default), `rpg`, `videogame`, `boardgameperson`, `rpgperson`,
`boardgamecompany`, `rpgcompany` and `videogamecompany`.}
}
\value{
a data frame containing nowadays' hot items.
}
\description{
Hot items nowadays
}
\examples{
hot_items_df <- hot()
hot_items_df
}
