% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanics.R
\name{mechanics}
\alias{mechanics}
\title{Mechanics of Board Games}
\usage{
mechanics()
}
\value{
Data frame containing all mechanics of board games (as per BGG)
}
\description{
\code{mechanics()} retrieves the board game mechanics that are
  present at \href{http://boardgamegeek.com}{Board Game Geek}.
}
\examples{
bg_mechanics <- mechanics()
bg_mechanics
}
\seealso{
\code{\link{categories}} \code{\link{designers}}
  \code{\link{families}}
}
