% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanics.R
\name{mechanics}
\alias{mechanics}
\title{Mechanics of board games}
\usage{
mechanics()
}
\value{
a data frame containing all mechanics of board games.
}
\description{
Mechanics of board games
}
\examples{
bg_mechanics <- mechanics()
bg_mechanics
}
\seealso{
[categories()]
}
