% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_k_games_ids.R
\name{top_k_games_ids}
\alias{top_k_games_ids}
\title{Top K Games IDs}
\usage{
top_k_games_ids(k = 100, wait = 10)
}
\arguments{
\item{k}{the number of top games that you want IDs for}

\item{wait}{number of seconds to wait between HTML pages as they are scraped}
}
\value{
Games IDs for the top k games
}
\description{
Top K Games IDs
}
\examples{
top_500_game_ids <- top_k_games_ids(k = 500)
top_500_game_ids
}
\seealso{
\code{\link{year_games_ids}}
}
