% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user}
\alias{user}
\title{User Info}
\usage{
user(username, buddies = 0, guilds = 0, hot = 0, top = 0)
}
\arguments{
\item{username}{name of the user to retrieve their details}

\item{buddies}{result to include user's buddies?}

\item{guilds}{result to include user's guilds?}

\item{hot}{result to include user's hot items?}

\item{top}{result to include user's top items?}
}
\value{
List containing the specified user's info.
}
\description{
\code{user()} retrieves the details for a specified user from
  \href{http://boardgamegeek.com}{Board Game Geek}.
}
\examples{
user_info <- user('alizat')
user_info
}
\seealso{
\code{\link{collection}}
}
