% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_games_ids.R
\name{year_games_ids}
\alias{year_games_ids}
\title{Specific Year's Games IDs}
\usage{
year_games_ids(y, wait = 5)
}
\arguments{
\item{y}{year to get games ids for}

\item{wait}{number of seconds to wait between HTML pages as they are scraped}
}
\value{
Games IDs for the top k games
}
\description{
Specific Year's Games IDs
}
\examples{
year_1995_game_ids <- year_games_ids(y = 1995)
year_1995_game_ids
}
\seealso{
\code{\link{top_k_games_ids}}
}
